backend/requirements.txt
fastapi
uvicorn
groq
gtts
python-dotenv
python-multipart
